% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time_utm_zone.R
\name{changeTZ}
\alias{changeTZ}
\title{Convert POSIXct objects to a new timezone}
\usage{
changeTZ(x, newTZ)
}
\arguments{
\item{x}{a POSIXct object}

\item{newTZ}{a valid timezone}
}
\value{
a POSIXct objects with datetime in newTZ timezone
}
\description{
\code{changeTZ} change the timezone of a POSIXct objects
}
\examples{
changeTZ(x,"Africa/Harare")
}

