% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_raw_data}
\alias{clean_raw_data}
\title{DAta clneaning title}
\usage{
clean_raw_data(input_file, output_file, timezone = "Etc/UTC",
  speed_limit = 110, acceleration_limit = 10)
}
\arguments{
\item{input_file}{numeric vector for passenger cars}

\item{output_file}{numeric vector for light commercial vehicles}

\item{timezone}{numeric vector for heavy good vehicles or trucks}

\item{speed_limit}{numeric vector for bus}

\item{acceleration_limit}{numeric vector for motorcycles}
}
\value{
data
}
\description{
\code{\link{clean_raw_data}} test lorem impo.
}
\examples{
\dontrun{
#do not run
}
}
