% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time_utm_zone.R
\name{CRS_UTM_zone}
\alias{CRS_UTM_zone}
\title{Generate CRS string for the UTM zone corresponding to latitude-longitude
coordinates}
\usage{
CRS_UTM_zone(latitude, longitude)
}
\arguments{
\item{latitude}{the latitude of location}

\item{longitude}{the longitude of location}
}
\value{
a CRS string
}
\description{
\code{CRS_UTM_zone} finds the UTM timezone corresponding to the longitude
locations given and generates the CRS string corresponding. Modified from
Cyril Bernard.
}
\examples{
CRS_UTM_zone(45,12)
}
