% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_tools.R
\name{extractPeriods}
\alias{extractPeriods}
\title{Extract periods in dateTime sequences}
\usage{
extractPeriods(x, sep = 7, unit = "days", id = NULL, burst = NULL)
}
\arguments{
\item{x}{POSIXct objects, a sequence of dateTime}

\item{sep}{maximum time of holes in date before introducing a new burst}

\item{unit}{unit of sep. Default "days"}

\item{id}{name of the sequence id}

\item{burst}{name of the sequence burst}
}
\value{
a dataframe with all periods and their date of begin (startPeriod) and end (endPeriod)
}
\description{
\code{extractPeriods} extracts the different sequences of observations in a dateTime sequence.
}
\examples{
extractPeriods(x)
}
\author{
Simon Chamaille-Jammes, \email{simon.chamaille@cefe.cnrs.fr}
}

