% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time_utm_zone.R
\name{FormatDateTime}
\alias{FormatDateTime}
\title{Format Date Time character string to POSIXct}
\usage{
FormatDateTime(x, TZ = "Etc/GMT")
}
\arguments{
\item{x}{a character vector}

\item{TZ}{the timezone of the object. Default GMT}
}
\value{
a POSIXct object
}
\description{
\code{FormatDateTime} automatically format Date Time objects into POSIXct. It assumes the dateTime character string can take a limited number of format : "%d/%m/%Y %H:%M:%S" ; "%m/%d/%Y %H:%M:%S" ; "%m/%d/%Y %I:%M:%S %p" ; "%d/%m/%Y %I:%M:%S %p"
}
\examples{
FormatDateTime(x)
}

