% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time_utm_zone.R
\name{get_julian}
\alias{get_julian}
\title{Get Julian Day}
\usage{
get_julian(x, origin = NULL)
}
\arguments{
\item{x}{a POSIXct object}

\item{origin}{facultative date object for origin of julian days. Default January, 1st 2000.}
}
\value{
a numeric object with days passed since origin
}
\description{
\code{get_julian} calculate julian day of a POSIXct object.
}
\examples{
changeTZ(x,"Africa/Harare")
}

