% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time_utm_zone.R
\name{get_tz}
\alias{get_tz}
\title{Find the generic timezone of longitude coordinates}
\usage{
get_tz(longitude)
}
\arguments{
\item{longitude}{the longitude of location}
}
\value{
a string with corresponding timezone. Beware of sign inversion :
  "Etc/GMT+2" is in reality GMT-2
}
\description{
\code{get_tz} finds the GMT+X timezone corresponding to the longitude
locations given.
}
\examples{
get_tz(150)
}

