% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_tools.R
\name{GetLetter}
\alias{GetLetter}
\title{Generate a sequence of letters}
\usage{
GetLetter(n)
}
\arguments{
\item{n}{number of element in the characters sequence}
}
\value{
a vector of 3 letters characters
}
\description{
\code{GetLetter} generates a sequence of 3 letter characters with a maximum
of 17576 elements
}
\examples{
GetLetter(n)
}

