% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpstools.R
\docType{package}
\name{gpstools}
\alias{gpstools}
\alias{gpstools-package}
\title{gpstools: tools for analysis of animal GPS movement data}
\description{
gpstools provides usefull tools for the analysis of animal movement such as
resampling trajectories with irregular data, creating timelines for
summarizing data, ...
}
