% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R, R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{clean_raw_data}
\title{Data cleaning}
\usage{
clean_raw_data(input_file, output_file, tmp_dir, timezone = "Etc/UTC",
  speed_limit = 110, acceleration_limit = 10, date_min = NA,
  date_max = NA, cores = 7)
}
\arguments{
\item{input_file}{file with the gps data to be cleaned}

\item{output_file}{file to write the output of the processing}

\item{tmp_dir}{where to place uncompressed input file}

\item{timezone}{represents the timezone}

\item{speed_limit}{numeric which contains the speed threshold (km/h)}

\item{acceleration_limit}{numeric which contains the acceleration threshold(m/s^2)}

\item{date_min}{Select data after this date (format:YYYY-MM-AA)}

\item{date_max}{Select data before this date (format:YYYY-MM-AA)}

\item{cores}{number of cores to be used}

\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\value{
data
}
\description{
\code{\link{clean_raw_data}} test lorem impo.

Like dplyr, ggvis also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
\dontrun{
#do not run
}
# Instead of
layer_points(ggvis(mtcars, ~mpg, ~wt))
# you can write
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
}
