% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_tools.R
\name{printday}
\alias{printday}
\title{Transform a number of days into Years, Month, Day}
\usage{
printday(n)
}
\arguments{
\item{n}{number of days}
}
\value{
a character
}
\description{
\code{printday} Transform a number of days into Years, Month, Day
}
\examples{
printday(450)
}

