% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_class.R
\name{timeline}
\alias{plot.timeline}
\alias{print.timeline}
\alias{timeline}
\title{timeline class description}
\usage{
\method{print}{timeline}(x)

\method{plot}{timeline}(x, theme = NULL)
}
\arguments{
\item{x}{a \code{timeline} object generated for instance by
\code{\link{create_timeline}}}
}
\description{
timeline class description

Print object of \code{timeline} class

Plot object of \code{timeline} class
}

