% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SIG_layer_functions.R
\name{write_shapefiles}
\alias{write_shapefiles}
\title{Write points and lines shapefiles for movement data/}
\usage{
write_shapefiles(df, dir, KML = FALSE, coord.x1 = "x", coord.x2 = "y",
  proj4string = "+proj=longlat +datum=WGS84 +no_defs", burst = "id",
  verbose = F)
}
\arguments{
\item{df}{data.frame with movement data}

\item{dir}{directory for writing the files}

\item{KML}{if true also write KML files into a KML subdirectory}

\item{coord.x1}{first coordinates}

\item{coord.x2}{second coordinates}

\item{proj4string}{CRS projection string for coordinates}

\item{burst}{colnames for separating the different files to write}

\item{verbose}{if true print burst progress}
}
\description{
\code{write_shapefiles} takes a dataframe with movement data from several
animals, splits it into the requested burst and write 2 shapefiles for each
burst. One with points and one with lines.
}
\examples{
write_shapefiles(df,"SIG/")
}

