% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_function.R
\name{timeline}
\alias{timeline}
\title{Creates a timeline for summarizing movement database}
\usage{
timeline(df, burstcol = "burstname", idcol = "id", timecol = "expectTime",
  color = NULL, actdf = NULL, acttimecol = "dateTime", sep = 7,
  unit = "days")
}
\arguments{
\item{df}{a data frame with data and a column in POSIXct}

\item{burstcol}{column names with defined burst}

\item{idcol}{column names for id}

\item{timecol}{column names for dateTime}

\item{color}{column names for the colors of data. If NULL no colors are added.}

\item{actdf}{a dataframe (optionnal) for adding activity duration. Must contains idcol and burstcol.}

\item{acttimecol}{colname for dateTime in actdf}

\item{sep}{maximum time between two locations to consider them in the same sequence}

\item{unit}{time unit of sep}
}
\value{
a \code{\link{timeline}}
}
\description{
\code{timeline} summarize for each individuals the duration of data available.
}
\examples{
data("hwzebra")
timeline(hwzebra)
}

